refactor: move direction package
