+++ src/main/java/com/holub/life/model/Direction.java
@@ -1,4 +1,4 @@
-package com.holub.life;
+package com.holub.life.model;
 
 import java.util.Objects;
