+++ src/main/java/com/holub/life/model/cell/DummyCell.java
@@ -1,6 +1,6 @@
 package com.holub.life.model.cell;
 
-import com.holub.life.Direction;
+import com.holub.life.model.Direction;
 import com.holub.tools.Observer;
 import com.holub.tools.Storable;
