+++ src/main/java/com/holub/life/model/cell/Neighborhood.java
@@ -1,7 +1,7 @@
 package com.holub.life.model.cell;
 
 import com.holub.asynch.ConditionVariable;
-import com.holub.life.Direction;
+import com.holub.life.model.Direction;
 import com.holub.life.model.Point;
 import com.holub.tools.Observer;
 import com.holub.tools.Storable;
